[Refactor] Refactor FTA related enums - rename EventType to FtaEventType and move to model.fta package - move GateType to model.fta package
