+++ src/main/java/cz/cvut/kbss/analysis/model/FaultEvent.java
@@ -1,13 +1,12 @@
 package cz.cvut.kbss.analysis.model;
 
-import cz.cvut.kbss.analysis.model.util.EventType;
-import cz.cvut.kbss.analysis.model.util.GateType;
+import cz.cvut.kbss.analysis.model.fta.FtaEventType;
+import cz.cvut.kbss.analysis.model.fta.GateType;
 import cz.cvut.kbss.analysis.model.diagram.Rectangle;
 import cz.cvut.kbss.analysis.util.Vocabulary;
 import cz.cvut.kbss.jopa.model.annotations.*;
 import lombok.Getter;
 import lombok.Setter;
-import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import java.net.URI;
 import java.util.*;
@@ -20,7 +19,7 @@ public class FaultEvent extends Event {
     @NotNull(message = "EventType must be defined")
     @ParticipationConstraints(nonEmpty = true)
     @OWLDataProperty(iri = Vocabulary.s_p_hasFaultEventType)
-    private EventType eventType;
+    private FtaEventType eventType;
 
     @OWLDataProperty(iri = Vocabulary.s_p_hasGateType)
     private GateType gateType;
