+++ src/main/java/cz/cvut/kbss/analysis/model/fta/FtaEventType.java
@@ -1,6 +1,6 @@
-package cz.cvut.kbss.analysis.model.util;
+package cz.cvut.kbss.analysis.model.fta;
 
-public enum EventType {
+public enum FtaEventType {
     BASIC,
     EXTERNAL,
     UNDEVELOPED,
